pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "beyond-prompting-repro"
version = "0.1.0"
description = "Local runnable reproduction package for 'Beyond Prompting: An Autonomous Framework for Systematic Factor Investing via Agentic AI'"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.1",
  "pyyaml>=6.0",
  "matplotlib>=3.8",
  "lightgbm>=4.3",
  "scikit-learn>=1.4",
  "statsmodels>=0.14",
  "pyarrow>=15.0"
]

[project.scripts]
agentic-factor = "agentic_factor.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["agentic_factor*"]
